Added entity for users attribute in reservation

+++ TutoringSystem/TutoringSystemLib/Entities/Subject.cs
@@ -8,7 +8,7 @@ namespace TutoringSystemLib.Entities
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal HourlRate  { get; set; }
+        public double HourlRate  { get; set; }
 
         public int LessonId { get; set; }
         public virtual Lesson Lesson { get; set; }
